refact(plugins): each plugin in a separate pkg
