+++ cmd/artoo/main.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/tfaughnan/artoo/client"
 	"github.com/tfaughnan/artoo/config"
-	"github.com/tfaughnan/artoo/plugin"
+	"github.com/tfaughnan/artoo/plugin/echo"
+	"github.com/tfaughnan/artoo/plugin/openai"
 )
 
 func main() {
@@ -19,8 +20,8 @@ func main() {
 	c.RegisterLineHandler(`^:(?P<server>\S+) 001 (?P<nick>\S+) :(?P<body>.*)$`, c.Handle001)
 	c.RegisterLineHandler(`^PING (?P<token>\S+)$`, c.HandlePing)
 	c.RegisterLineHandler(`^:(?P<nick>\S+)!(?P<user>\S+)@(?P<host>\S+) PRIVMSG (?P<target>\S+) :(?P<body>.*)$`, c.HandlePrivmsg)
-	c.RegisterPluginHandler(plugin.EchoPattern, plugin.EchoHandler)
-	c.RegisterPluginHandler(plugin.OpenaiPattern, plugin.OpenaiHandler)
+	c.RegisterPluginHandler(echo.EchoPattern, echo.EchoHandler)
+	c.RegisterPluginHandler(openai.OpenaiPattern, openai.OpenaiHandler)
 
 	if err := c.Connect(); err != nil {
 		log.Fatal(err)
